+++ packages/typespec-ts/src/modular/type-expressions/get-type-expression.ts
@@ -63,7 +63,6 @@ export function getTypeExpression(
       }
       return "number";
     case "endpoint":
-    case "plainDate":
     case "plainTime":
     case "string":
     case "url":
@@ -104,6 +103,7 @@ export function getTypeExpression(
     case "union":
       return getUnionExpression(context, type, options);
     case "utcDateTime":
+    case "plainDate":
       return "Date";
 
     default: